requirements.txt
# Core Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# YOLOv8 for Segmentation
ultralytics>=8.0.0

# Segment Anything Model (SAM) for auto-labeling
segment-anything
git+https://github.com/facebookresearch/segment-anything.git

# Computer Vision
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
Pillow>=9.5.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Metrics and Evaluation
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Progress Bars
tqdm>=4.65.0

# YAML Configuration
PyYAML>=6.0

# Logging
tensorboard>=2.13.0
wandb  # Optional: for experiment tracking

# Utilities
albumentations>=1.3.0  # Data augmentation
timm>=0.9.0  # Model architectures

# LLM Assistance Tracking (课程要求)
# 本项目使用了 GitHub Copilot 辅助生成代码
# 具体标注见各文件开头注释
